moved enable logic to core range facet
https://coveord.atlassian.net/browse/KIT-1421
