+++ packages/headless/src/controllers/core/facets/range-facet/date-facet/headless-core-date-facet.ts
@@ -39,11 +39,6 @@ import {
 import {loadReducerError} from '../../../../../utils/errors';
 import {deselectAllFacetValues} from '../../../../../features/facets/facet-set/facet-set-actions';
 import {CoreEngine} from '../../../../../app/engine';
-import {isFacetEnabledSelector} from '../../../../../features/facet-options/facet-options-selectors';
-import {
-  enableFacet,
-  disableFacet,
-} from '../../../../../features/facet-options/facet-options-actions';
 
 export type {
   DateFacetOptions,
@@ -182,8 +177,6 @@ export function buildCoreDateFacet(
 
   validateDateFacetOptions(engine, options);
 
-  const getIsEnabled = () => isFacetEnabledSelector(engine.state, facetId);
-
   dispatch(registerDateFacet(options));
 
   const rangeFacet = buildCoreRangeFacet<DateFacetRequest, DateFacetResponse>(
@@ -208,21 +201,8 @@ export function buildCoreDateFacet(
       this.toggleSelect(selection);
     },
 
-    enable() {
-      dispatch(enableFacet(facetId));
-    },
-
-    disable() {
-      dispatch(disableFacet(facetId));
-    },
-
     get state() {
-      const enabled = getIsEnabled();
-
-      return {
-        ...rangeFacet.state,
-        enabled,
-      };
+      return rangeFacet.state;
     },
   };
 }
